helloworld/requirements.txt
Flask
flask-assets
redis
pymongo
pyyaml